Move `create_csr_matrix_from_count_df`
Moves this function into the only module it's actually used in. This
function was too specific to `CountVectorizer` to make sense outside of
this module.
